+++ pangeo_forge_recipes/patterns.py
@@ -5,6 +5,7 @@ import inspect
 import warnings
 from dataclasses import dataclass, field, replace
 from enum import Enum, auto
+from hashlib import sha256
 from itertools import product
 from typing import (
     Any,
@@ -20,7 +21,7 @@ from typing import (
     Union,
 )
 
-from .serialization import pattern_blockchain
+from .serialization import dataclass_sha256, dict_drop_empty, dict_to_sha256
 
 
 class CombineOp(Enum):
@@ -295,3 +296,64 @@ def prune_pattern(fp: FilePattern, nkeep: int = 2) -> FilePattern:
         if param not in ["format_function", "combine_dims"]
     }
     return FilePattern(fp.format_function, *new_combine_dims, **kwargs)
+
+
+def pattern_blockchain(pattern: FilePattern) -> List[bytes]:
+    """For a ``FilePattern`` instance, compute a blockchain, i.e. a list of hashes of length N+1,
+    where N is the number of index:filepath pairs yielded by the ``FilePattern`` instance's
+    ``.items()`` method. The first item in the list is calculated by hashing instance attributes.
+    Each subsequent item is calculated by hashing the byte string produced by concatenating the next
+    index:filepath pair yielded by ``.items()`` with the previous hash in the list.
+
+    :param pattern: The ``FilePattern`` instance for which to calculate a blockchain.
+    """
+
+    # we exclude the format function and combine dims from ``root`` because they determine the
+    # index:filepath pairs yielded by iterating over ``.items()``. if these pairs are generated in
+    # a different way in the future, we ultimately don't care.
+    root = {
+        "fsspec_open_kwargs": pattern.fsspec_open_kwargs,
+        "query_string_secrets": pattern.query_string_secrets,
+        "file_type": pattern.file_type,
+        "nitems_per_file": [
+            op.nitems_per_file  # type: ignore
+            for op in pattern.combine_dims
+            if op.name in pattern.concat_dims
+        ],
+    }
+    # by dropping empty values from ``root``, we allow for the attributes of ``FilePattern`` to
+    # change while allowing for backwards-compatibility between hashes of patterns which do not
+    # set those new attributes.
+    root_drop_empty = dict_drop_empty([(k, v) for k, v in root.items()])
+    root_sha256 = dict_to_sha256(root_drop_empty)
+
+    blockchain = [root_sha256]
+    for k, v in pattern.items():
+        key_hash = b"".join(
+            sorted([dataclass_sha256(dimindex, ignore_keys=["sequence_len"]) for dimindex in k])
+        )
+        value_hash = sha256(v.encode("utf-8")).digest()
+        new_hash = key_hash + value_hash
+        new_block = sha256(blockchain[-1] + new_hash).digest()
+        blockchain.append(new_block)
+
+    return blockchain
+
+
+def match_pattern_blockchain(  # type: ignore
+    old_pattern_last_hash: bytes,
+    new_pattern: FilePattern,
+) -> Index:
+    """Given the last hash of the blockchain for a previous pattern, and a new pattern, determine
+    which (if any) ``Index`` key of the new pattern to begin processing data from, in order to
+    append to a dataset build using the previous pattern.
+
+    :param old_pattern_last_hash: The last hash of the blockchain for the ``FilePattern`` instance
+      which was used to build the existing dataset.
+    :param new_pattern: A new ``FilePattern`` instance from which to append to the existing dataset.
+    """
+
+    new_chain = pattern_blockchain(new_pattern)
+    for k, h in zip(new_pattern, new_chain):
+        if h == old_pattern_last_hash:
+            return k
